refactor(test): 프로덕트 코드에 맞게 테스트 파일 패키지 구조 변경
