+++ src/test/java/domain/participant/DealerTest.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.participant;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -6,7 +6,6 @@ import domain.card.Card;
 import domain.card.Cards;
 import domain.card.Denomination;
 import domain.card.Suit;
-import domain.participant.Dealer;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -18,7 +17,8 @@ public class DealerTest {
     void canAddCardTrue() {
         //given 
         Cards cards = new Cards(
-            List.of(new Card(Suit.CLOVER, Denomination.FIVE), new Card(Suit.CLOVER, Denomination.SIX)));
+            List.of(new Card(Suit.CLOVER, Denomination.FIVE),
+                new Card(Suit.CLOVER, Denomination.SIX)));
 
         //when
         Dealer dealer = new Dealer(cards);
@@ -32,7 +32,8 @@ public class DealerTest {
     void canAddCardFalse() {
         //given
         Cards cards = new Cards(
-            List.of(new Card(Suit.CLOVER, Denomination.A), new Card(Suit.CLOVER, Denomination.SIX)));
+            List.of(new Card(Suit.CLOVER, Denomination.A),
+                new Card(Suit.CLOVER, Denomination.SIX)));
 
         //when
         Dealer dealer = new Dealer(cards);
